Refactor core and cdi
